backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.4
torch==2.0.0
torchvision==0.15.0
numpy==1.24.3
opencv-python==4.7.0
easyocr==1.6.2
tensorflow==2.12.0
Pillow==9.5.0
ensemble-boxes==1.0.9
